renamed imagemanifestvuln db model to dashdotdb
