+++ .gitignore
@@ -2,3 +2,5 @@ venv/
 .idea/
 __pycache__
 *.egg-info/
+build/
+dist/
